+++ docs/CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for Sentry error reporting. Set `SENTRY_DSN` at compile- or run-time to enable reporting.
-- Added Prometheus counters: `log_warn_count`, `log_error_count` and `log_critical_count` representing the corresponding number of warning/error/critical messages in the log.
+- Added Prometheus counters: `log_warn_count`, `log_error_count`, `log_critical_count`, `log_panic_count` and `log_fatal_count` representing the corresponding number of warning/error/critical/panic/fatal messages in the log.
 
 #### New locking mode: 'lease'